Added a setting to allow Amazon S3 hosted images to work in Chromium browsers
Amazon S3 has some strange behaviour with cors/non-cors caching in
chromium based browsers. To get around this we can add  "?
cacheblock=true" to the image URL being fetched, this stops the item
(and most importantly in the case the headers) being cached.  There is
a possible performance issue here I guess but it seems pretty minimal
as it is just getting the image again from S3.  With this modification S3
can be used happily.

Thread about the issue:
https://bugs.chromium.org/p/chromium/issues/detail?id=409090